backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
openai==1.51.0
orjson==3.10.7
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
